Add files via upload
BH_Perturbation_report1 corresponds to the work done in phase 1 of this project. In phase 1, concepts of GR were revised with setting up the theoretical framework for understanding black hole dynamics and different types of perturbations in it.

Naxxatra_Final_Report summarises the final work done in this project. Different graphs have been added related to QNMs of different pertubations in a black hole and their trends with different parameres like mass, radius and angular momentum parameter (l).